LLM_GRIEVANCE_TICKET_GENERATION/requirements.txt
langchain==0.1.13
streamlit==1.29.0
openai==1.14.2
tiktoken==0.5.2
python-dotenv==1.0.1
pinecone-client==3.2.0
PyPDF2
joblib==1.3.2
pandas==2.2.0
scikit-learn==1.4.0
sentence-transformers==2.5.1
langchain-openai==0.1.0